Add timeouts for compilation and execution
The limits are set to be quite generous, so it shouldn't affect anything that isn't already broken, especially since all the test cases are rather small. In particular, if a test case hangs in the current one, the only way to kill it is sigint or similar, and then you don't get to see what case failed. With timeouts, it can run through the full suite, though it will take a while.

No timeout is set for building the compiler itself though, as that has legit reasons to take a while.